CSC4101 Prog 1 - Initial Code - DS
This is what I've done so far. The Scanner and Parser work for the most
part. Parser builds a correct tree that can be traversed by retrieiving
the car and cdr values from the node all the way on down. I think the
ordering should be correct.

The Scanner is not completely done. The DOT type has to be worked out
and I may have bugs when it comes to Strings (when there is not closed
string). Also need to teach it to skip comments.

There's no real error checking anywhere in it.